replace "request" with "http" and "ws" stream handlers
